pyproject.toml
[project]
name = "rlm-plus-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.24",
    "anthropic>=0.75.0",
    "groq>=1.0.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
]
